issue #15: add required YUI3 node-event-simulate module
When feedback comments are enabled, the `core_filepicker` module is loaded. This module requires and loads several YUI3 modules to the page, including `node-event-simulate` that the error message refers to.

The list of these can be found in `/lib/outputrequirementslib.php` in the
`find_module` function that is called whenever a JS module is
initialised.

When feedback comments are disabled, the `node-event-simulate` is not
loaded. This commit adds this YUI3 module when loading the plugin's JS
module.